refactor: 파일명 config에서 constants로 변경
